Slideshow functionalities
Trying to make the slideshow work, it currently changes the current image to the next and if the last image is selected it will go back to the first. The only problem I am having is the previous and next slideshow image option, I am trying to figure out how to change the value of the image but when I call the changeImg function the if statement negates the changes of the previousSlideClick function.